+++ back/moya/src/main/java/com/study/moya/chat/text/controller/TextChatRoomController.java
@@ -7,6 +7,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,7 +21,8 @@ public class TextChatRoomController {
     private final ChatService chatService;
 
     @GetMapping
-    public ResponseEntity<List<ChatRoomDTO>> getAllRooms() {
+    public ResponseEntity<List<ChatRoomDTO>> getUserRooms(Authentication authentication) {
+        log.info("get user rooms");
         List<ChatRoomDTO> rooms = chatService.findAllRooms();
         return ResponseEntity.ok(rooms);
     }
@@ -30,8 +32,7 @@ public class TextChatRoomController {
         String userEmail = principal.getName();
         log.debug("채팅방 생성 요청 - 사용자: {}, 방 이름: {}", userEmail, roomName);
 
-        ChatRoomDTO chatRoom = chatService.createTextRoom(roomName);
-        // 방 생성자를 첫 번째 참여자로 추가
+        ChatRoomDTO chatRoom = chatService.createTextRoom(roomName, userEmail);
         chatService.addUserToRoom(chatRoom.getRoomId(), userEmail);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(chatRoom);
@@ -64,18 +65,16 @@ public class TextChatRoomController {
         }
     }
 
-    // 채팅방 삭제 엔드포인트 추가
     @DeleteMapping("/room/{roomId}")
     public ResponseEntity<?> deleteRoom(@PathVariable String roomId, Principal principal) {
         String userEmail = principal.getName();
         log.debug("채팅방 삭제 요청 - 사용자: {}, 방 ID: {}", userEmail, roomId);
 
         try {
-//            ChatRoomDTO room = chatService.findRoomById(roomId);
-//            // 방장만 삭제할 수 있도록 체크
-//            if (!room.getCreator().equals(userEmail)) {
-//                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-//            }
+            ChatRoomDTO room = chatService.findRoomById(roomId);
+            if (!room.getCreator().equals(userEmail)) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
 
             chatService.deleteRoom(roomId);
             return ResponseEntity.ok().build();
